CORE-2388 Migrate BranchPostTask from standalone class back to inner class of Branch object
