+++ Branch-SDK/src/main/java/io/branch/referral/Branch.java
@@ -17,7 +17,6 @@ import android.content.pm.PackageManager;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.AsyncTask;
-import android.os.Build;
 import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -26,6 +25,7 @@ import androidx.annotation.StyleRes;
 import android.os.Handler;
 import android.os.Looper;
 import android.text.TextUtils;
+import android.util.Log;
 import android.view.View;
 
 import androidx.core.app.ActivityCompat;
@@ -75,6 +75,9 @@ import io.branch.referral.util.LinkProperties;
  * -->
  */
 public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserver.AdsParamsFetchEvents, GooglePlayStoreAttribution.IInstallReferrerEvents {
+
+    //Sanity check for AAR inclusion:
+    public static final int SANITY_CHECK = 3;
     
     private static final String BRANCH_LIBRARY_VERSION = "io.branch.sdk.android:library:" + Branch.getSdkVersionNumber();
     private static final String GOOGLE_VERSION_TAG = "!SDK-VERSION-STRING!" + ":" + BRANCH_LIBRARY_VERSION;
@@ -1685,7 +1688,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
 
     private void executeTimedBranchPostTask(final ServerRequest req, final int timeout) {
         final CountDownLatch latch = new CountDownLatch(1);
-        final BranchPostTask postTask = new BranchPostTask(branchReferral_, req, latch);
+        final BranchPostTask postTask = new BranchPostTask(req, latch);
 
         postTask.executeTask();
         if (Looper.myLooper() == Looper.getMainLooper()) {
@@ -2233,6 +2236,184 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
             return branchRemoteInterface_.make_restful_post(serverRequests[0].getPost(),
                     prefHelper_.getAPIBaseUrl() + Defines.RequestPath.GetURL.getPath(),
                     Defines.RequestPath.GetURL.getPath(), prefHelper_.getBranchKey());
+        }
+    }
+
+    /**
+     * Asynchronous task handling execution of server requests. Execute the network task on background
+     * thread and request are  executed in sequential manner. Handles the request execution in
+     * Synchronous-Asynchronous pattern. Should be invoked only form main thread and  the results are
+     * published in the main thread.
+     */
+    private class BranchPostTask extends BranchAsyncTask<Void, Void, ServerResponse> {
+        ServerRequest thisReq_;
+        final CountDownLatch latch_;
+
+        public BranchPostTask(ServerRequest request, CountDownLatch latch) {
+            super();
+            thisReq_ = request;
+            latch_ = latch;
+        }
+
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+            thisReq_.onPreExecute();
+            thisReq_.doFinalUpdateOnMainThread();
+        }
+
+        @Override
+        protected ServerResponse doInBackground(Void... voids) {
+            // update queue wait time
+            addExtraInstrumentationData(thisReq_.getRequestPath() + "-" + Defines.Jsonkey.Queue_Wait_Time.getKey(), String.valueOf(thisReq_.getQueueWaitTime()));
+            thisReq_.doFinalUpdateOnBackgroundThread();
+            if (isTrackingDisabled() && !thisReq_.prepareExecuteWithoutTracking()) {
+                return new ServerResponse(thisReq_.getRequestPath(), BranchError.ERR_BRANCH_TRACKING_DISABLED, "");
+            }
+            String branchKey = prefHelper_.getBranchKey();
+            ServerResponse result;
+            if (thisReq_.isGetRequest()) {
+                result = getBranchRemoteInterface().make_restful_get(thisReq_.getRequestUrl(), thisReq_.getGetParams(), thisReq_.getRequestPath(), branchKey);
+            } else {
+                result = getBranchRemoteInterface().make_restful_post(thisReq_.getPostWithInstrumentationValues(instrumentationExtraData_), thisReq_.getRequestUrl(), thisReq_.getRequestPath(), branchKey);
+            }
+            if (latch_ != null) {
+                latch_.countDown();
+            }
+            return result;
+        }
+
+        @Override
+        protected void onPostExecute(ServerResponse serverResponse) {
+            super.onPostExecute(serverResponse);
+            onPostExecuteInner(serverResponse);
+        }
+
+        void onPostExecuteInner(ServerResponse serverResponse) {
+            if (latch_ != null) {
+                latch_.countDown();
+            }
+            if (serverResponse == null) {
+                thisReq_.handleFailure(BranchError.ERR_BRANCH_INVALID_REQUEST, "Null response.");
+                return;
+            }
+
+            int status = serverResponse.getStatusCode();
+            if (status == 200) {
+                onRequestSuccess(serverResponse);
+            } else {
+                onRequestFailed(serverResponse, status);
+            }
+            networkCount_ = 0;
+
+            processNextQueueItem();
+        }
+
+        private void onRequestSuccess(ServerResponse serverResponse) {
+            // If the request succeeded
+            @Nullable final JSONObject respJson = serverResponse.getObject();
+            if (respJson == null) {
+                thisReq_.handleFailure(500, "Null response json.");
+            }
+
+            if (thisReq_ instanceof ServerRequestCreateUrl && respJson != null) {
+                try {
+                    // cache the link
+                    BranchLinkData postBody = ((ServerRequestCreateUrl) thisReq_).getLinkPost();
+                    final String url = respJson.getString("url");
+                    linkCache_.put(postBody, url);
+                } catch (JSONException ex) {
+                    ex.printStackTrace();
+                }
+            } else if (thisReq_ instanceof ServerRequestLogout) {
+                //On Logout clear the link cache and all pending requests
+                linkCache_.clear();
+                requestQueue_.clear();
+            }
+
+
+            if (thisReq_ instanceof ServerRequestInitSession || thisReq_ instanceof ServerRequestIdentifyUserRequest) {
+                // If this request changes a session update the session-id to queued requests.
+                boolean updateRequestsInQueue = false;
+                if (!isTrackingDisabled() && respJson != null) {
+                    // Update PII data only if tracking is disabled
+                    try {
+                        if (respJson.has(Defines.Jsonkey.SessionID.getKey())) {
+                            prefHelper_.setSessionID(respJson.getString(Defines.Jsonkey.SessionID.getKey()));
+                            updateRequestsInQueue = true;
+                        }
+                        if (respJson.has(Defines.Jsonkey.IdentityID.getKey())) {
+                            String new_Identity_Id = respJson.getString(Defines.Jsonkey.IdentityID.getKey());
+                            if (!prefHelper_.getIdentityID().equals(new_Identity_Id)) {
+                                //On setting a new identity Id clear the link cache
+                                linkCache_.clear();
+                                prefHelper_.setIdentityID(new_Identity_Id);
+                                updateRequestsInQueue = true;
+                            }
+                        }
+                        if (respJson.has(Defines.Jsonkey.DeviceFingerprintID.getKey())) {
+                            prefHelper_.setDeviceFingerPrintID(respJson.getString(Defines.Jsonkey.DeviceFingerprintID.getKey()));
+                            updateRequestsInQueue = true;
+                        }
+                        if (updateRequestsInQueue) {
+                            updateAllRequestsInQueue();
+                        }
+                    } catch (JSONException ex) {
+                        ex.printStackTrace();
+                    }
+                }
+
+                if (thisReq_ instanceof ServerRequestInitSession) {
+                    setInitState(Branch.SESSION_STATE.INITIALISED);
+                    if (!((ServerRequestInitSession) thisReq_).handleBranchViewIfAvailable((serverResponse))) {
+                        checkForAutoDeepLinkConfiguration();
+                    }
+                    // Count down the latch holding getLatestReferringParamsSync
+                    if (getLatestReferringParamsLatch != null) {
+                        getLatestReferringParamsLatch.countDown();
+                    }
+                    // Count down the latch holding getFirstReferringParamsSync
+                    if (getFirstReferringParamsLatch != null) {
+                        getFirstReferringParamsLatch.countDown();
+                    }
+                }
+            }
+
+            if (respJson != null) {
+                thisReq_.onRequestSucceeded(serverResponse, branchReferral_);
+                requestQueue_.remove(thisReq_);
+            } else if (thisReq_.shouldRetryOnFail()) {
+                // already called handleFailure above
+                thisReq_.clearCallbacks();
+            } else {
+                requestQueue_.remove(thisReq_);
+            }
+        }
+
+        void onRequestFailed(ServerResponse serverResponse, int status) {
+            // If failed request is an initialisation request (but not in the intra-app linking scenario) then mark session as not initialised
+            if (thisReq_ instanceof ServerRequestInitSession && PrefHelper.NO_STRING_VALUE.equals(prefHelper_.getSessionParams())) {
+                setInitState(Branch.SESSION_STATE.UNINITIALISED);
+            }
+
+            // On a bad request or in case of a conflict notify with call back and remove the request.
+            if ((status == 400 || status == 409) && thisReq_ instanceof ServerRequestCreateUrl) {
+                ((ServerRequestCreateUrl) thisReq_).handleDuplicateURLError();
+            } else {
+                //On Network error or Branch is down fail all the pending requests in the queue except
+                //for request which need to be replayed on failure.
+                networkCount_ = 0;
+                thisReq_.handleFailure(status, serverResponse.getFailReason());
+            }
+
+            boolean unretryableErrorCode = (400 <= status && status <= 451);
+            if (unretryableErrorCode || !thisReq_.shouldRetryOnFail()) {
+                requestQueue_.remove(thisReq_);
+            } else {
+                // failure has already been handled
+                // todo does it make sense to retry the request without a callback? (e.g. CPID, LATD)
+                thisReq_.clearCallbacks();
+            }
         }
     }
